tests/[TEST UNITAIRE] - lambda_docker_upload_one_chart/requirements.txt
kaleido==0.1.*
pandas
plotly
boto3
psycopg2-binary
